pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent"
version = "0.1.0"
description = "AI agent for transcription, data analysis and workflow intelligence"
authors=[{name="Rafid Mahbub"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.9",
    "celery>=5.6.2",
    "ddgs>=9.11.2",
    "fastapi>=0.133.1",
    "faster-whisper>=1.2.1",
    "ffmpeg-python>=0.2.0",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "itsdangerous>=2.2.0",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-openai>=1.1.10",
    "langchain-postgres>=0.0.17",
    "langgraph>=1.0.9",
    "langgraph-checkpoint-postgres>=3.0.4",
    "langsmith[openai-agents]>=0.7.7",
    "onnxruntime>=1.19.2",
    "openai>=2.24.0",
    "pandas>=2.3.3",
    "pandas-stubs~=2.3.3",
    "plotly>=6.5.2",
    "psycopg2>=2.9.11",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "pypdf>=6.7.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "redis>=7.2.1",
    "sqlmodel>=0.0.37",
    "starlette>=0.52.1",
    "streamlit>=1.55.0",
    "torch>=2.10.0",
    "uvicorn>=0.41.0",
    "watchdog>=6.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
